Update the EVM how-to guide (#172)
I'm trying to update the EVM docs to work as a how-to guide, meaning there are specific steps that you can follow to use real-time pyth data.

Please LMK what you think of the new format. We can fine-tune it here and then propagate to other parts of the docs.