fix: re-enable SSR
To be quite honest, I'm not sure if this is safe or what it might break (or why
this was ever done in the first place).  However, having SSR disabled is causing
some docs pages to have terrible performance and people are complaining.

I did some testing with this and I didn't see any issues.  At this point I
believe the best path forward is to re-enable SSR and just fix forward if we
find any issues I missed